requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30

torch>=2.3
torchvision>=0.18
pillow>=10.2

python-multipart>=0.0.9